# lxdweb/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Session 的认证",
        "title": "LXD Web 管理平台",
        "termsOfService": "https://github.com/xkatld/zjmf-lxd-server",
        "contact": {
            "name": "xkatld",
            "url": "https://github.com/xkatld/zjmf-lxd-server",
            "email": "xkatld@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE"
        },
        "version": "1.0.1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/nodes": {
            "get": {
                "description": "查询所有LXD节点及其系统信息，支持按状态过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "节点管理"
                ],
                "summary": "获取节点列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点状态(active/inactive)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回节点列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "添加新的LXD节点到管理系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "节点管理"
                ],
                "summary": "创建节点",
                "parameters": [
                    {
                        "description": "节点配置参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateNodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/nodes/{id}": {
            "get": {
                "description": "根据ID获取节点详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "节点管理"
                ],
                "summary": "获取单个节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回节点信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新节点配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "节点管理"
                ],
                "summary": "更新节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateNodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "从管理系统中删除指定节点及其所有关联数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "节点管理"
                ],
                "summary": "删除节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/nodes/{id}/refresh": {
            "post": {
                "description": "刷新指定节点的系统信息缓存",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "节点管理"
                ],
                "summary": "刷新节点缓存",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/nodes/{id}/test": {
            "post": {
                "description": "测试与LXD节点的连接状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "节点管理"
                ],
                "summary": "测试节点连接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "测试成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "连接失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateNodeRequest": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "api_key": {
                    "type": "string"
                },
                "batch_interval": {
                    "type": "integer"
                },
                "batch_size": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sync_preset": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateNodeRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "api_key": {
                    "type": "string"
                },
                "batch_interval": {
                    "type": "integer"
                },
                "batch_size": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sync_preset": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}